LPS-73024 Moves to portal-test-integration to reuse it on modules
