+++ modules/apps/web-experience/trash/trash-test/src/testIntegration/java/com/liferay/trash/service/webserver/test/WebServerTrashTest.java
@@ -44,6 +44,7 @@ import com.liferay.portal.kernel.webdav.methods.Method;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.test.randomizerbumpers.TikaSafeRandomizerBumper;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.webserver.test.BaseWebServerTestCase;
 import com.liferay.portlet.documentlibrary.service.permission.DLPermission;
 import com.liferay.trash.kernel.model.TrashEntry;
