+++ portal-impl/test/integration/com/liferay/portal/webserver/WebServerRangeTest.java
@@ -29,6 +29,7 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.webdav.methods.Method;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.webserver.test.BaseWebServerTestCase;
 
 import java.util.Arrays;
 import java.util.HashMap;
