+++ portal-test-integration/src/com/liferay/portal/webserver/test/BaseWebServerTestCase.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.webserver;
+package com.liferay.portal.webserver.test;
 
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.servlet.HttpHeaders;
@@ -20,7 +20,8 @@ import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portlet.documentlibrary.service.BaseDLAppTestCase;
+import com.liferay.portal.webserver.WebServerServlet;
+import com.liferay.portlet.documentlibrary.service.test.BaseDLAppTestCase;
 
 import java.util.HashMap;
 import java.util.Map;
